commands: refactored new project creation
